backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# AI and ML
google-genai
faster-whisper

# Audio processing
sounddevice
scipy
pyttsx3
gTTS
comtypes
pywin32

# Utilities
python-dotenv
